requirements.txt
llama-index-core>=0.12.0
llama-index-readers-file>=0.4.0
llama-index-embeddings-openai>=0.3.0
llama-index-llms-openai>=0.3.0
faiss-cpu>=1.9.0
streamlit>=1.40.0
openai>=1.50.0
python-dotenv>=1.0.0
pydantic>=2.0
pytest>=8.0
